--- frontend/src/components/Flex/index.ts
@@ -1,1 +1,0 @@
-export { default as Flex } from './Flex';
