+++ frontend/src/components/Header/Header.tsx
@@ -2,10 +2,7 @@ import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import logoIcon from '@/assets/images/logo.png';
 import newTemplateIcon from '@/assets/images/newTemplate.png';
-import { Button } from '../Button';
-import { Flex } from '../Flex';
-import { Input } from '../Input';
-import { Text } from '../Text';
+import { Button, Flex, Input, Text } from '..';
 import * as S from './style';
 
 const Header = () => {
