+++ frontend/src/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import { Header } from '../Header';
+import { Header } from '..';
 import * as S from './style';
 
 const Layout = () => (
