--- frontend/src/components/SelectList/index.ts
@@ -1,1 +1,0 @@
-export { default as SelectList } from './SelectList';
