--- frontend/src/components/SnippetEditor/index.ts
@@ -1,1 +1,0 @@
-export { default as SnippetEditor } from './SnippetEditor';
