+++ frontend/src/components/TemplateItem/TemplateItem.tsx
@@ -2,8 +2,7 @@ import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
 import { vscDarkPlus } from 'react-syntax-highlighter/dist/esm/styles/prism';
 import { TemplateListItem } from '@/types/template';
 import { formatRelativeTime } from '@/utils/formatRelativeTime';
-import { Flex } from '../Flex';
-import { Text } from '../Text';
+import { Flex, Text } from '..';
 
 interface Props {
   item: TemplateListItem;
