--- frontend/src/components/TemplateItem/index.ts
@@ -1,1 +1,0 @@
-export { default as TemplateItem } from './TemplateItem';
