--- frontend/src/components/TemplateTitleInput/index.ts
@@ -1,1 +1,0 @@
-export { default as TemplateTitleInput } from './TemplateTitleInput';
