--- frontend/src/components/Text/index.ts
@@ -1,1 +1,0 @@
-export { default as Text } from './Text';
