+++ frontend/src/components/index.ts
@@ -1,0 +1,10 @@
+export { default as Button } from './Button/Button';
+export { default as Flex } from './Flex/Flex';
+export { default as Header } from './Header/Header';
+export { default as Input } from './Input/Input';
+export { default as Layout } from './Layout/Layout';
+export { default as SelectList } from './SelectList/SelectList';
+export { default as SnippetEditor } from './SnippetEditor/SnippetEditor';
+export { default as TemplateItem } from './TemplateItem/TemplateItem';
+export { default as TemplateTitleInput } from './TemplateTitleInput/TemplateTitleInput';
+export { default as Text }from './Text/Text';
