+++ frontend/src/pages/Template.tsx
@@ -2,9 +2,7 @@ import { useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
 import { vscDarkPlus } from 'react-syntax-highlighter/dist/esm/styles/prism';
-import { Flex } from '@/components/Flex';
-import { SelectList } from '@/components/SelectList';
-import { Text } from '@/components/Text';
+import { Flex, SelectList, Text } from '@/components';
 import { useTemplateQuery } from '@/hooks/useTemplateQuery';
 import { formatRelativeTime } from '@/utils/formatRelativeTime';
