+++ frontend/src/pages/TemplateList.tsx
@@ -1,7 +1,5 @@
 import { Link } from 'react-router-dom';
-import { Flex } from '@/components/Flex';
-import { TemplateItem } from '@/components/TemplateItem';
-import { Text } from '@/components/Text';
+import { Flex, TemplateItem, Text } from '@/components';
 import { useTemplateListQuery } from '@/hooks/useTemplateListQuery';
 
 const TemplateList = () => {
